LP64 Variant 3 sweep: bare -1 through untyped dispatch, AMS window-size defaults + trymail cwp_Search
FossilOrigin-Name: 20d50ccc0a50f73b4a30f0baef5b5c078c53ab6af6f758216db60d973013d5d9

+++ src/atkams/messages/lib/messwind.c
@@ -66,7 +66,7 @@ struct messwind *self;
 {
     struct messwind *m = (struct messwind *) messwind_New();
 
-    ams_InstallInNewWindow(m, "messages", "Messages", environ_GetProfileInt("messages.width", -1), environ_GetProfileInt("messages.height", -1), m->folders);
+    ams_InstallInNewWindow(m, "messages", "Messages", environ_GetProfileInt("messages.width", (long)-1), environ_GetProfileInt("messages.height", (long)-1), m->folders);
 }
 
 static struct keymap *messwind_standardkeymap;
